Move log classes to log dir
